% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint_resources.R
\name{CheckpointResourceClient}
\alias{CheckpointResourceClient}
\title{A verified public checkpoint resource}
\description{
A verified public checkpoint resource

A verified public checkpoint resource
}
\details{
Resource credentials are discarded after immediate acquisition. Generic
table conversion is deliberately unavailable.
}
\section{Super class}{
\code{\link[resourcer:ResourceClient]{resourcer::ResourceClient}} -> \code{CheckpointResourceClient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CheckpointResourceClient-new}{\code{CheckpointResourceClient$new()}}
\item \href{#method-CheckpointResourceClient-getSnapshot}{\code{CheckpointResourceClient$getSnapshot()}}
\item \href{#method-CheckpointResourceClient-asDataFrame}{\code{CheckpointResourceClient$asDataFrame()}}
\item \href{#method-CheckpointResourceClient-asTbl}{\code{CheckpointResourceClient$asTbl()}}
\item \href{#method-CheckpointResourceClient-clone}{\code{CheckpointResourceClient$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="close"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-close'><code>resourcer::ResourceClient$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="downloadFile"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-downloadFile'><code>resourcer::ResourceClient$downloadFile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="exec"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-exec'><code>resourcer::ResourceClient$exec()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="getConnection"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-getConnection'><code>resourcer::ResourceClient$getConnection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="getResource"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-getResource'><code>resourcer::ResourceClient$getResource()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckpointResourceClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-CheckpointResourceClient-new}{}}}
\subsection{Method \code{new()}}{
Acquire and verify a registered descriptor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckpointResourceClient$new(resource)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{A custodian-registered resource descriptor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckpointResourceClient-getSnapshot"></a>}}
\if{latex}{\out{\hypertarget{method-CheckpointResourceClient-getSnapshot}{}}}
\subsection{Method \code{getSnapshot()}}{
Return verified node-private snapshot state to trusted R code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckpointResourceClient$getSnapshot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckpointResourceClient-asDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-CheckpointResourceClient-asDataFrame}{}}}
\subsection{Method \code{asDataFrame()}}{
Refuse generic dataset conversion.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckpointResourceClient$asDataFrame(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckpointResourceClient-asTbl"></a>}}
\if{latex}{\out{\hypertarget{method-CheckpointResourceClient-asTbl}{}}}
\subsection{Method \code{asTbl()}}{
Refuse generic dataset conversion.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckpointResourceClient$asTbl(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckpointResourceClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CheckpointResourceClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckpointResourceClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
