% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkpoint_resources.R
\name{CheckpointResourceResolver}
\alias{CheckpointResourceResolver}
\title{Resolver for custodian-registered checkpoint bundles}
\description{
Resolver for custodian-registered checkpoint bundles

Resolver for custodian-registered checkpoint bundles
}
\section{Super class}{
\code{\link[resourcer:ResourceResolver]{resourcer::ResourceResolver}} -> \code{CheckpointResourceResolver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CheckpointResourceResolver-isFor}{\code{CheckpointResourceResolver$isFor()}}
\item \href{#method-CheckpointResourceResolver-newClient}{\code{CheckpointResourceResolver$newClient()}}
\item \href{#method-CheckpointResourceResolver-clone}{\code{CheckpointResourceResolver$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceResolver" data-id="initialize"><a href='../../resourcer/html/ResourceResolver.html#method-ResourceResolver-initialize'><code>resourcer::ResourceResolver$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckpointResourceResolver-isFor"></a>}}
\if{latex}{\out{\hypertarget{method-CheckpointResourceResolver-isFor}{}}}
\subsection{Method \code{isFor()}}{
Match only the digest-pinned dsFlower checkpoint format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckpointResourceResolver$isFor(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Resource descriptor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckpointResourceResolver-newClient"></a>}}
\if{latex}{\out{\hypertarget{method-CheckpointResourceResolver-newClient}{}}}
\subsection{Method \code{newClient()}}{
Materialize a verified checkpoint snapshot immediately.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckpointResourceResolver$newClient(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Resource descriptor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CheckpointResourceResolver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CheckpointResourceResolver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CheckpointResourceResolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
