% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{.apply_privacy_reservation}
\alias{.apply_privacy_reservation}
\title{Bind an accountant reservation to a prepared manifest}
\usage{
.apply_privacy_reservation(staging_dir, reservation)
}
\description{
Bind an accountant reservation to a prepared manifest
}
\keyword{internal}
