% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{.canonicalPatientIds}
\alias{.canonicalPatientIds}
\title{Canonicalise and validate a complete patient roster}
\usage{
.canonicalPatientIds(values)
}
\description{
Canonicalise and validate a complete patient roster
}
\keyword{internal}
