% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{.excludePatientFeature}
\alias{.excludePatientFeature}
\title{Remove the protected patient identifier from model features}
\usage{
.excludePatientFeature(feature_columns, patient_column)
}
\description{
Remove the protected patient identifier from model features
}
\keyword{internal}
