% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{.getHandle}
\alias{.getHandle}
\title{Retrieve authoritative state for an opaque Flower handle}
\usage{
.getHandle(symbol)
}
\description{
Retrieve authoritative state for an opaque Flower handle
}
\keyword{internal}
