% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{.install_integrity_hook}
\alias{.install_integrity_hook}
\title{Install the mandatory Python code-integrity bootstrap}
\usage{
.install_integrity_hook(
  staging_dir,
  hook_src = system.file("python", "sitecustomize.py", package = "dsFlower")
)
}
\description{
Install the mandatory Python code-integrity bootstrap
}
\keyword{internal}
