% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_ledger.R
\name{.ledger_path}
\alias{.ledger_path}
\title{Get the path to the privacy ledger file}
\usage{
.ledger_path()
}
\value{
Character; path to the ledger JSON file.
}
\description{
Get the path to the privacy ledger file
}
\keyword{internal}
