% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{.new_handle_capability}
\alias{.new_handle_capability}
\title{Generate an opaque Flower-handle capability}
\usage{
.new_handle_capability()
}
\description{
Generate an opaque Flower-handle capability
}
\keyword{internal}
