% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.node_secret_path}
\alias{.node_secret_path}
\title{Resolve the dedicated dsFlower node-secret path}
\usage{
.node_secret_path()
}
\description{
Resolve the dedicated dsFlower node-secret path
}
\keyword{internal}
