% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{.overlay_download_tailscale}
\alias{.overlay_download_tailscale}
\title{Download the static tailscale binaries (tailscaled + tailscale) on demand}
\usage{
.overlay_download_tailscale()
}
\description{
Download the static tailscale binaries (tailscaled + tailscale) on demand
}
\keyword{internal}
