% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{.overlay_start_forward}
\alias{.overlay_start_forward}
\title{Start a SOCKS5 forwarder: loopback port -> tailnet SuperLink address}
\usage{
.overlay_start_forward(target_addr)
}
\description{
Start a SOCKS5 forwarder: loopback port -> tailnet SuperLink address
}
\keyword{internal}
