% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_env.R
\name{.python_deps_for_framework}
\alias{.python_deps_for_framework}
\title{Get all pip dependencies for a framework}
\usage{
.python_deps_for_framework(framework)
}
\description{
Get all pip dependencies for a framework
}
\keyword{internal}
