% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_ledger.R
\name{.read_ledger}
\alias{.read_ledger}
\title{Read the privacy ledger}
\usage{
.read_ledger()
}
\value{
Named list of dataset entries with RDP vectors.
}
\description{
Read the privacy ledger
}
\keyword{internal}
