% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_env.R
\name{.read_venv_python_version}
\alias{.read_venv_python_version}
\title{Read the Python version recorded by a virtual environment}
\usage{
.read_venv_python_version(venv_path)
}
\description{
Read the Python version recorded by a virtual environment
}
\keyword{internal}
