% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{.removeHandle}
\alias{.removeHandle}
\title{Remove authoritative state for an opaque Flower handle}
\usage{
.removeHandle(symbol, cleanup = TRUE)
}
\description{
Remove authoritative state for an opaque Flower handle
}
\keyword{internal}
