% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staging.R
\name{.stageFromDescriptor_df}
\alias{.stageFromDescriptor_df}
\title{Stage from an in-memory data.frame descriptor}
\usage{
.stageFromDescriptor_df(
  desc,
  run_token,
  target_column,
  feature_columns,
  extra_config
)
}
\description{
Stage from an in-memory data.frame descriptor
}
\keyword{internal}
