% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tor.R
\name{.tor_ensure}
\alias{.tor_ensure}
\title{Ensure Tor is running with a SOCKS5 proxy on this node}
\usage{
.tor_ensure()
}
\description{
Ensure Tor is running with a SOCKS5 proxy on this node
}
\keyword{internal}
