% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsi_tunnel.R
\name{.tunnel_generation}
\alias{.tunnel_generation}
\title{Validate a tunnel connection generation}
\usage{
.tunnel_generation(value)
}
\description{
Validate a tunnel connection generation
}
\keyword{internal}
