% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsi_tunnel.R
\name{.tunnel_publish_ack}
\alias{.tunnel_publish_ack}
\title{Publish one monotonic absolute tunnel acknowledgement}
\usage{
.tunnel_publish_ack(spool, name, offset)
}
\description{
Publish one monotonic absolute tunnel acknowledgement
}
\keyword{internal}
