% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsi_tunnel.R
\name{.tunnel_reset_spool_file}
\alias{.tunnel_reset_spool_file}
\title{Replace one spool with an empty stream starting at an absolute offset}
\usage{
.tunnel_reset_spool_file(spool, binname, base = 0)
}
\description{
Replace one spool with an empty stream starting at an absolute offset
}
\keyword{internal}
