% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsi_tunnel.R
\name{.tunnel_spool_max_bytes}
\alias{.tunnel_spool_max_bytes}
\title{Maximum bytes retained in either tunnel spool file}
\usage{
.tunnel_spool_max_bytes(chunk_bytes = .tunnel_chunk_bytes())
}
\description{
Maximum bytes retained in either tunnel spool file
}
\keyword{internal}
