% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{.validate_handle_capability}
\alias{.validate_handle_capability}
\title{Validate an opaque Flower-handle capability}
\usage{
.validate_handle_capability(capability)
}
\description{
Validate an opaque Flower-handle capability
}
\keyword{internal}
