% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.validate_node_secret}
\alias{.validate_node_secret}
\title{Validate a dedicated 256-bit dsFlower node secret}
\usage{
.validate_node_secret(path)
}
\description{
Validate a dedicated 256-bit dsFlower node secret
}
\keyword{internal}
