% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.validate_node_secret_parent}
\alias{.validate_node_secret_parent}
\title{Validate the parent directory of a dsFlower node secret}
\usage{
.validate_node_secret_parent(path, euid = NULL)
}
\description{
Validate the parent directory of a dsFlower node secret
}
\keyword{internal}
