% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_env.R
\name{.venv_is_healthy}
\alias{.venv_is_healthy}
\title{Check if a Python venv is healthy}
\usage{
.venv_is_healthy(venv_path, framework)
}
\description{
Check if a Python venv is healthy
}
\keyword{internal}
