% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_env.R
\name{.venv_site_packages_dirs}
\alias{.venv_site_packages_dirs}
\title{Locate site-packages directories inside a virtual environment}
\usage{
.venv_site_packages_dirs(venv_path)
}
\description{
Locate site-packages directories inside a virtual environment
}
\keyword{internal}
