% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsi_tunnel.R
\name{.with_tunnel_lock}
\alias{.with_tunnel_lock}
\title{Serialize aggregate exchanges belonging to one tunnel session}
\usage{
.with_tunnel_lock(spool, code)
}
\description{
Serialize aggregate exchanges belonging to one tunnel session
}
\keyword{internal}
