% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_ledger.R
\name{.write_ledger}
\alias{.write_ledger}
\title{Write the privacy ledger}
\usage{
.write_ledger(ledger)
}
\arguments{
\item{ledger}{Named list (the ledger data).}
}
\value{
Invisible NULL.
}
\description{
Write the privacy ledger
}
\keyword{internal}
