% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.write_node_secret_atomic}
\alias{.write_node_secret_atomic}
\title{Atomically write a fresh dsFlower node secret}
\usage{
.write_node_secret_atomic(path, parent)
}
\description{
Atomically write a fresh dsFlower node secret
}
\keyword{internal}
