% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_store.R
\name{flowerAppDeleteDS}
\alias{flowerAppDeleteDS}
\title{Remove an uploaded app's spool (DataSHIELD AGGREGATE)}
\usage{
flowerAppDeleteDS(token)
}
\arguments{
\item{token}{Character; upload token.}
}
\value{
TRUE.
}
\description{
Remove an uploaded app's spool (DataSHIELD AGGREGATE)
}
\keyword{internal}
