% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{flowerListTemplatesDS}
\alias{flowerListTemplatesDS}
\title{List Available Templates}
\usage{
flowerListTemplatesDS()
}
\value{
Character vector of template names.
}
\description{
DataSHIELD AGGREGATE method. Returns the names of all Flower app
templates installed on this server.
}
