% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{flowerOverlayDownDS}
\alias{flowerOverlayDownDS}
\title{Tear down the overlay on this node}
\usage{
flowerOverlayDownDS()
}
\value{
Invisible TRUE.
}
\description{
DataSHIELD ASSIGN method. Stops the forwarder and tailscaled.
}
