% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{FlowerResourceClient}
\alias{FlowerResourceClient}
\title{Flower Federation Resource Client}
\description{
Flower Federation Resource Client

Flower Federation Resource Client
}
\details{
R6 class that wraps a DataSHIELD resource pointing to a Flower federation
node. Extracts configuration from the resource URL.
}
\section{URL Format}{

\code{flower+federation:///data_path=/data/train.csv;data_format=csv;python_path=python3}
}

\keyword{internal}
\section{Super class}{
\code{\link[resourcer:ResourceClient]{resourcer::ResourceClient}} -> \code{FlowerResourceClient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FlowerResourceClient-new}{\code{FlowerResourceClient$new()}}
\item \href{#method-FlowerResourceClient-getParsed}{\code{FlowerResourceClient$getParsed()}}
\item \href{#method-FlowerResourceClient-getDataPath}{\code{FlowerResourceClient$getDataPath()}}
\item \href{#method-FlowerResourceClient-getDataFormat}{\code{FlowerResourceClient$getDataFormat()}}
\item \href{#method-FlowerResourceClient-getPythonPath}{\code{FlowerResourceClient$getPythonPath()}}
\item \href{#method-FlowerResourceClient-close}{\code{FlowerResourceClient$close()}}
\item \href{#method-FlowerResourceClient-clone}{\code{FlowerResourceClient$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asDataFrame"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-asDataFrame'><code>resourcer::ResourceClient$asDataFrame()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asTbl"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-asTbl'><code>resourcer::ResourceClient$asTbl()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="downloadFile"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-downloadFile'><code>resourcer::ResourceClient$downloadFile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="exec"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-exec'><code>resourcer::ResourceClient$exec()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="getConnection"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-getConnection'><code>resourcer::ResourceClient$getConnection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="getResource"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-getResource'><code>resourcer::ResourceClient$getResource()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowerResourceClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-FlowerResourceClient-new}{}}}
\subsection{Method \code{new()}}{
Create a new Flower resource client
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowerResourceClient$new(resource, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{A resourcer resource object}

\item{\code{...}}{Additional arguments passed to parent}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowerResourceClient-getParsed"></a>}}
\if{latex}{\out{\hypertarget{method-FlowerResourceClient-getParsed}{}}}
\subsection{Method \code{getParsed()}}{
Get parsed URL parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowerResourceClient$getParsed()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Named list with all parsed parameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowerResourceClient-getDataPath"></a>}}
\if{latex}{\out{\hypertarget{method-FlowerResourceClient-getDataPath}{}}}
\subsection{Method \code{getDataPath()}}{
Get the training data path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowerResourceClient$getDataPath()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string or NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowerResourceClient-getDataFormat"></a>}}
\if{latex}{\out{\hypertarget{method-FlowerResourceClient-getDataFormat}{}}}
\subsection{Method \code{getDataFormat()}}{
Get the data format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowerResourceClient$getDataFormat()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string (csv, parquet, feather)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowerResourceClient-getPythonPath"></a>}}
\if{latex}{\out{\hypertarget{method-FlowerResourceClient-getPythonPath}{}}}
\subsection{Method \code{getPythonPath()}}{
Get the Python executable path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowerResourceClient$getPythonPath()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowerResourceClient-close"></a>}}
\if{latex}{\out{\hypertarget{method-FlowerResourceClient-close}{}}}
\subsection{Method \code{close()}}{
Close the resource client (no-op for federation)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowerResourceClient$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowerResourceClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FlowerResourceClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowerResourceClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
