% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{FlowerResourceResolver}
\alias{FlowerResourceResolver}
\title{Flower Federation Resource Resolver}
\description{
Flower Federation Resource Resolver

Flower Federation Resource Resolver
}
\details{
A \code{resourcer::ResourceResolver} subclass that creates Flower
federation clients from DataSHIELD resource descriptors.
}
\keyword{internal}
\section{Super class}{
\code{\link[resourcer:ResourceResolver]{resourcer::ResourceResolver}} -> \code{FlowerResourceResolver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FlowerResourceResolver-isFor}{\code{FlowerResourceResolver$isFor()}}
\item \href{#method-FlowerResourceResolver-newClient}{\code{FlowerResourceResolver$newClient()}}
\item \href{#method-FlowerResourceResolver-clone}{\code{FlowerResourceResolver$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceResolver" data-id="initialize"><a href='../../resourcer/html/ResourceResolver.html#method-ResourceResolver-initialize'><code>resourcer::ResourceResolver$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowerResourceResolver-isFor"></a>}}
\if{latex}{\out{\hypertarget{method-FlowerResourceResolver-isFor}{}}}
\subsection{Method \code{isFor()}}{
Check if this resolver can handle the given resource
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowerResourceResolver$isFor(resource)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{A resourcer resource object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowerResourceResolver-newClient"></a>}}
\if{latex}{\out{\hypertarget{method-FlowerResourceResolver-newClient}{}}}
\subsection{Method \code{newClient()}}{
Create a new client for the given resource
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowerResourceResolver$newClient(resource)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{A resourcer resource object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A FlowerResourceClient, or NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlowerResourceResolver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FlowerResourceResolver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlowerResourceResolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
