% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsi_tunnel.R
\name{flowerTunnelDownDS}
\alias{flowerTunnelDownDS}
\title{Stop this tunnel session's forwarder and clean its spool (AGGREGATE)}
\usage{
flowerTunnelDownDS(conn_id)
}
\description{
Stop this tunnel session's forwarder and clean its spool (AGGREGATE)
}
\keyword{internal}
