% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsi_tunnel.R
\name{flowerTunnelInjectDS}
\alias{flowerTunnelInjectDS}
\title{TEST: inject bytes as if the SuperNode sent them (DataSHIELD AGGREGATE)}
\usage{
flowerTunnelInjectDS(conn_id, data_b64)
}
\arguments{
\item{conn_id}{Character; tunnel connection id.}

\item{data_b64}{Character; B64: url-safe payload to append to up.bin.}
}
\value{
TRUE.
}
\description{
TEST: inject bytes as if the SuperNode sent them (DataSHIELD AGGREGATE)
}
\keyword{internal}
