% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsi_tunnel.R
\name{flowerTunnelLogDS}
\alias{flowerTunnelLogDS}
\title{Read a tunnel-related log tail (DataSHIELD AGGREGATE)}
\usage{
flowerTunnelLogDS(conn_id, which = "sn")
}
\arguments{
\item{conn_id}{Character; tunnel id.}

\item{which}{One of "sn" (supernode), "fwd" (forwarder), "tc" (test client).}
}
\value{
Character; last lines of the log.
}
\description{
Read a tunnel-related log tail (DataSHIELD AGGREGATE)
}
\keyword{internal}
