% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsi_tunnel.R
\name{flowerTunnelPushDS}
\alias{flowerTunnelPushDS}
\title{Deliver SuperLink->SuperNode bytes from the relay (DataSHIELD AGGREGATE)}
\usage{
flowerTunnelPushDS(conn_id, data_b64)
}
\arguments{
\item{conn_id}{Character; tunnel connection id.}

\item{data_b64}{Character; B64: url-safe payload to append to down.bin.}
}
\value{
TRUE.
}
\description{
Deliver SuperLink->SuperNode bytes from the relay (DataSHIELD AGGREGATE)
}
\keyword{internal}
