% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsi_tunnel.R
\name{flowerTunnelResetDS}
\alias{flowerTunnelResetDS}
\title{Reset/initialise a tunnel spool (DataSHIELD AGGREGATE)}
\usage{
flowerTunnelResetDS(conn_id)
}
\arguments{
\item{conn_id}{Character; tunnel connection id.}
}
\value{
TRUE.
}
\description{
Reset/initialise a tunnel spool (DataSHIELD AGGREGATE)
}
\keyword{internal}
