% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsi_tunnel.R
\name{flowerTunnelTestClientDS}
\alias{flowerTunnelTestClientDS}
\title{TEST: spawn a local client that round-trips through the forwarder (AGGREGATE)}
\usage{
flowerTunnelTestClientDS(
  conn_id,
  listen_port,
  msg = "HELLO-TUNNEL",
  nrep = 1000L
)
}
\description{
TEST: spawn a local client that round-trips through the forwarder (AGGREGATE)
}
\keyword{internal}
