% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsi_tunnel.R
\name{flowerTunnelTestResultDS}
\alias{flowerTunnelTestResultDS}
\title{TEST: read the round-tripped bytes the client received (AGGREGATE)}
\usage{
flowerTunnelTestResultDS(conn_id)
}
\description{
TEST: read the round-tripped bytes the client received (AGGREGATE)
}
\keyword{internal}
