% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptor.R
\name{print.FlowerDatasetDescriptor}
\alias{print.FlowerDatasetDescriptor}
\title{Print method for FlowerDatasetDescriptor}
\usage{
\method{print}{FlowerDatasetDescriptor}(x, ...)
}
\arguments{
\item{x}{A FlowerDatasetDescriptor.}

\item{...}{Ignored.}
}
\description{
Print method for FlowerDatasetDescriptor
}
