% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{flowerDestroyDS}
\alias{flowerDestroyDS}
\title{Destroy Flower Handle}
\usage{
flowerDestroyDS(handle_symbol)
}
\arguments{
\item{handle_symbol}{Character; symbol of the handle.}
}
\value{
NULL.
}
\description{
DataSHIELD ASSIGN method. Full cleanup: removes staging, stops
the associated SuperNode, and removes the handle.
}
