% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{flowerLogDS}
\alias{flowerLogDS}
\title{Get Sanitized Log Output}
\usage{
flowerLogDS(handle_symbol, last_n = 50L)
}
\arguments{
\item{handle_symbol}{Character; symbol of the handle.}

\item{last_n}{Integer; number of lines to return (max 200).}
}
\value{
Character vector of sanitized log lines.
}
\description{
DataSHIELD AGGREGATE method. Returns sanitized log output from the
SuperNode.
}
