% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{flowerPrivacyBudgetDS}
\alias{flowerPrivacyBudgetDS}
\title{Query Privacy Budget}
\usage{
flowerPrivacyBudgetDS(handle_symbol)
}
\arguments{
\item{handle_symbol}{Character; symbol of the handle.}
}
\value{
Named list with spent/remaining epsilon and delta.
}
\description{
DataSHIELD AGGREGATE method. Returns the remaining privacy budget
for the dataset associated with the handle.
}
