% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsi_tunnel.R
\name{flowerTunnelDownDS}
\alias{flowerTunnelDownDS}
\title{Stop the node-side tunnel forwarder/test processes and clean the spool (AGGREGATE)}
\usage{
flowerTunnelDownDS(conn_id)
}
\description{
Stop the node-side tunnel forwarder/test processes and clean the spool (AGGREGATE)
}
\keyword{internal}
